Tue, Jun 7, 2022, 11:50 PM +00:00

+++ filter-less/helpers.c
@@ -175,9 +175,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 avgR += imageCopy[i - 1][j - 1].rgbtRed;
                 surroundingCount++;
             }
-            image[i][j].rgbtBlue = round(avgB / surroundingCount);
-            image[i][j].rgbtGreen = round(avgG / surroundingCount);
-            image[i][j].rgbtRed = round(avgR / surroundingCount);
+            image[i][j].rgbtBlue = round(avgB / (double)surroundingCount);
+            image[i][j].rgbtGreen = round(avgG / (double)surroundingCount);
+            image[i][j].rgbtRed = round(avgR / (double)surroundingCount);
         }
     }
     return;
